Started writing initial simulation source
Started writing a preliminary phase-field crystal simulation.
This included modifying the CMake files to be able to build, link, and
install the files.
Additionally, I modified the CMake files to correctly handle the
Runpaths, deal with Debug vs. Release mode, create compile_commands,
    etc.
I documented all of this behavior in the README so that hopefully
installation of the library is doable.