Started folder to hold phase_field_functions
On many occasions, we will have need for functions which describe a
particular phase field -- this will be necessary if we ever use
Dirichlet boundary conditions, but it will also be important when
initializing the phase field.
These functions will also necessarily initialize the auxiliary variables
(since they are finite element fields as well).