Initialized configuration in hexagonal lattice
Wrote a function which gives back the phase field and auxiliary
variables for a hexagonal lattice as a function of positions.
Included this function in the initialization of the fe field for the
phase field crystal system.
This took a lot of modifying dependencies, but it seems like it works
without crashing.
Next step is to actuall output the configuration to make sure everything
looks okay.