Set periodic boundaries, checked by outputting
Used functionality in deal.II to set periodic boundary conditions with
the AffineConstraints object.
Slightly displaced the HexagonalLattice function in order to check that
it was working properly (if the configuration is somewhat offset, then
        the edges don't line up in values, and so the imposed periodic
        boundary conditions make the configuration look a little bit
        weird in order to conform to the imposed boundary conditions).
The output from the vtu file looks wonky in the way that is expected, so
I think the imposed periodic conditions worked.