Completely removed params from class
Before I had been partially initializing some of the parameters for the
simulation from within the class, either as default arguments, or just
within random functions.
That's bad if I want to change them, because then I have to recompile
everything, and it's also just bad to have all of the parameters spread
throughout the class.
Hence, I took them out and they're all passed to the constructor.