Added ability to use toml files
I would like to specify parameters via toml files, and in order to do
that I used an external toml library.
This is included via the git submodules functionality.
Because it can be used as a header-only library, but I do not want it to
be (it slows down compilation times a lot) I had to set up an internal
library which basically links the functionality.
Given this, it is only compiled once in the toml.cpp compilation unit.
To use this library, you just include the internal parameters/toml.hpp
header file which links to the internal toml.cpp file.